added rrd memory mode "none" to allow netdata act as a proxy to forward metrics to another netdata
