+++ src/common.h
@@ -217,6 +217,8 @@
 #include "inlined.h"
 #include "adaptive_resortable_list.h"
 #include "rrdpush.h"
+#include "web_api_v1.h"
+#include "web_api_old.h"
 
 extern char *netdata_configured_config_dir;
 extern char *netdata_configured_log_dir;
